+++ test/subscriptions.tests.js
@@ -2176,6 +2176,103 @@ describe(
       );
     });
 
+    it('should limit concurrent messages using dynamic channel prefetch while subscribing', (test, done) => {
+      createBroker(
+        {
+          vhosts,
+          publications,
+          subscriptions: {
+            s1: {
+              vhost: '/',
+              queue: 'q1',
+              channelPrefetch: 1,
+            },
+          },
+        },
+        (err, broker) => {
+          assert.ifError(err);
+          async.times(
+            10,
+            (index, next) => {
+              broker.publish('p1', 'test message', next);
+            },
+            (err) => {
+              assert.ifError(err);
+              const messages = [];
+              broker.subscribe('s1', (err, subscription) => {
+                assert.ifError(err);
+                subscription.on('message', (message, content, ackOrNack) => {
+                  assert(message);
+                  messages.push(ackOrNack);
+                  if (messages.length === 5) {
+                    setTimeout(() => {
+                      assert.strictEqual(messages.length, 5);
+                      subscription.cancel(done);
+                      setTimeout(() => {
+                        messages.forEach((ackOrNack) => ackOrNack());
+                      }, 1);
+                    }, 500);
+                  }
+                  subscription.setChannelPrefetch(5, (err) => {
+                    assert.ifError(err);
+                  });
+                });
+              });
+            }
+          );
+        }
+      );
+    });
+
+    it('should limit concurrent messages using dynamic channel prefetch before subscription starts', (test, done) => {
+      createBroker(
+        {
+          vhosts,
+          publications,
+          subscriptions: {
+            s1: {
+              vhost: '/',
+              queue: 'q1',
+              channelPrefetch: 1,
+            },
+          },
+        },
+        (err, broker) => {
+          assert.ifError(err);
+          async.times(
+            10,
+            (index, next) => {
+              broker.publish('p1', 'test message', next);
+            },
+            (err) => {
+              assert.ifError(err);
+              const messages = [];
+              broker.subscribe('s1', (err, subscription) => {
+                assert.ifError(err);
+                subscription.setChannelPrefetch(5, (err) => {
+                  assert.ifError(err);
+
+                  subscription.on('message', (message, content, ackOrNack) => {
+                    assert(message);
+                    messages.push(ackOrNack);
+                    if (messages.length === 5) {
+                      setTimeout(() => {
+                        assert.strictEqual(messages.length, 5);
+                        subscription.cancel(done);
+                        setTimeout(() => {
+                          messages.forEach((ackOrNack) => ackOrNack());
+                        }, 1);
+                      }, 500);
+                    }
+                  });
+                });
+              });
+            }
+          );
+        }
+      );
+    });
+
     it('should emit channel errors', (test, done) => {
       createBroker(
         {